miq_observe - move all JS code to one place
no further changes yet
