+++ app/assets/javascripts/miq_application.js
@@ -1,4 +1,4 @@
-/* global add_flash dialogFieldRefresh getChartColumnDataValues getChartFormatedValue miqBrowserDetect miqExpressionPrefill miqFlashLater miqFlashSaved miqGridCheckAll miqGridGetCheckedRows miqLoadTL miqMenu miqTreeObject miqValueStylePrefill performFiltering recalculateChartYAxisLabels */
+/* global add_flash getChartColumnDataValues getChartFormatedValue miqBrowserDetect miqExpressionPrefill miqFlashLater miqFlashSaved miqGridCheckAll miqGridGetCheckedRows miqLoadTL miqMenu miqTreeObject miqValueStylePrefill performFiltering recalculateChartYAxisLabels */
 
 // MIQ specific JS functions
 
@@ -874,27 +874,6 @@ function miqBuildCalendar() {
       ManageIQ.observeDate = observeDateBackup;
     }
   });
-}
-
-function miqSendDateRequest(el) {
-  var parms = $.parseJSON(el.attr('data-miq_observe_date'));
-  var url = parms.url;
-  //  tack on the id and value to the URL
-  var urlstring = url + '?' + el.prop('id') + '=' + el.val();
-
-  var attemptAutoRefreshTrigger = function() {
-    if (parms.auto_refresh === true) {
-      dialogFieldRefresh.triggerAutoRefresh(parms);
-    }
-  };
-
-  var options = {
-    beforeSend: !!el.attr('data-miq_sparkle_on'),
-    complete: !!el.attr('data-miq_sparkle_off'),
-    done: attemptAutoRefreshTrigger,
-  };
-
-  return miqObserveRequest(urlstring, options);
 }
 
 // common function to pass ajax request to server
