+++ app/javascript/packs/application-common.js
@@ -21,6 +21,7 @@ import { initializeStore } from '../miq-redux';
 import { history } from '../miq-component/react-history.ts';
 import createReduxRoutingActions from '../miq-redux/redux-router-actions';
 import { formButtonsActionTypes, createFormButtonsActions } from '../forms/form-buttons-reducer';
+import { setup as miqObserveSetup } from '../miq_observe.js';
 
 ManageIQ.component = {
   ...newRegistry,
@@ -55,3 +56,6 @@ require('xml_display/XMLDisplay.css');
 
 // miqSpinner, miqSearchSpinner
 window.Spinner = Spinner;
+
+// used by miq_ujs_bindings.js
+window.miqObserveSetup = miqObserveSetup;
